template/requirements.txt
wheel==0.37.0
apache-beam[gcp]==2.32.0
Pillow==8.3.2
imageio==2.9.0